Add files via upload
File contains updated Anchored SPM poverty estimates from 2010 to 2020 for children under 18. This is a subset of a larger dataset that contains Anchored SPM poverty estimates from 1967 to 2020. See the original data set for more detail. The TT in the filename stands for "Taxes and Transfers". Taxes and transfers are policies, like pandemic stimulus checks, that have shown to positively impact (reduce) the rate of childhood poverty when implemented. 

Data is from:
https://www.povertycenter.columbia.edu/historical-spm-data
https://www.vox.com/22600143/poverty-us-covid-19-pandemic-stimulus-checks